add reset function and modal on existing_cluster page

+++ terraform/appflask.py
@@ -1367,12 +1367,24 @@ def read_process(g):
         for prcs, line in lines:
             yield line
 
+
+@app.route('/reset', methods=['POST'])
 def reset():
+    fabric_type = get_fabric_type(request)
     '''generic function to delete the TF State'''
-    if os.path.exists(TF_STATE_ACI):
-        os.remove(TF_STATE_ACI)
-    if os.path.exists(TF_STATE_NDFC):
-        os.remove(TF_STATE_NDFC)
+    if fabric_type not in VALID_FABRIC_TYPE:
+        return redirect('/')
+    if request.method == "POST":
+        try:
+            if fabric_type == "aci" and os.path.exists(TF_STATE_ACI):
+                os.remove(TF_STATE_ACI)
+            if fabric_type == "vxlan_evpn" and os.path.exists(TF_STATE_NDFC):
+                os.remove(TF_STATE_NDFC)
+        except Exception:
+            print("reset failed")
+            return json.dumps({"error": "reset failed"}), 503
+        return json.dumps({"msg": "reset success"}), 200
+
 
 @app.route('/existing_cluster', methods=['GET', 'POST'])
 def existing_cluster():
